Feature: Add extension client logger
Previously, the language server had a dedicated channel in the Output
window. Since the client will start getting more complex, it will also
require to log messages. This commit adds a logger class that will
handle logging to the Output window and the console.

The Output channel name for the language server was renamed to start
with "BitBake" to be easy to associate with the extension.