Fix: Change working directory for relative paths settings
Settings like pathToBitbakeFolder should use "${workspaceFolder}" to
point to relative URIs. However if the user uses a relative path like
"./", the language server did resolve it properly, but not the client.
We now set the working directory to the workspace folder to be
consistent.

This does not produce unintended side effects according to our tests.